pyproject.toml
[project]
name = "kindle-display-server"
version = "0.2.0"
description = "A Python-based server that generates custom grayscale images for Kindle e-ink displays"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "matplotlib>=3.8.2",
    "numpy>=1.26.2",
    "pillow>=10.1.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.2",
    "python-multipart>=0.0.6",
    "google-auth>=2.23.4",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.108.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.8",
    "pyright>=1.1.338",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B905",  # zip strict parameter (not available in Python 3.11)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pyright]
include = ["app", "tests"]
exclude = ["**/__pycache__", ".venv"]
typeCheckingMode = "basic"
reportMissingImports = true
reportMissingTypeStubs = false
pythonVersion = "3.11"
